backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
pdfplumber
python-docx
openpyxl
python-pptx
beautifulsoup4
requests
